+++ Source/Controllers/BundleSupport/SPBundleEditorController.m
@@ -1052,14 +1052,9 @@
 	[fileManager removeItemAtPath:cmdFilePath error:nil];
 	NSError *error = nil;
 
-	if (@available(macOS 10.13, *)) {
-		[saveDict writeToURL:[NSURL fileURLWithPath:cmdFilePath] error:&error];
-		if(error){
-			SPLog(@"Could not write %@. Error: %@", cmdFilePath, error.localizedDescription);
-		}
-
-	} else {
-		[saveDict writeToFile:cmdFilePath atomically:YES];
+	[saveDict writeToURL:[NSURL fileURLWithPath:cmdFilePath] error:&error];
+	if(error){
+		SPLog(@"Could not write %@. Error: %@", cmdFilePath, error.localizedDescription);
 	}
 
 	return YES;